+++ app/model_synchronizers/work_model_synchronizer.rb
@@ -1,0 +1,61 @@
+# frozen_string_literal: true
+
+# Updates a Work model object based on a DRO
+#
+# Syncing should be performed whenever the DRO is retrieved (since it may have been updated externally from H3)
+# and whenever the work is updated by the user.
+class WorkModelSynchronizer
+  class Error < StandardError; end
+
+  def self.call(...)
+    new(...).call
+  end
+
+  # @param [Work] work
+  # @param [Cocina::Models::DROWithMetadata,Cocina::Models::RequestDRO] cocina_object
+  # @param [Boolean] raise whether to raise an error if the collection is not found
+  # @raise [Error] if the collection is not found
+  def initialize(work:, cocina_object:, raise: true)
+    @work = work
+    @cocina_object = cocina_object
+    @raise = raise
+  end
+
+  def call
+    work.update!(**update_params)
+  end
+
+  private
+
+  attr_reader :work, :cocina_object
+
+  def update_params
+    {
+      title:,
+      collection:,
+      version:
+    }.tap do |params|
+      params[:object_updated_at] = cocina_object.try(:modified)
+    end.compact
+  end
+
+  def title
+    Cocina::Parser.title_for(cocina_object:)
+  end
+
+  def version
+    Cocina::Parser.version_for(cocina_object:)
+  end
+
+  def collection_druid
+    @collection_druid ||= Cocina::Parser.collection_druid_for(cocina_object:)
+  end
+
+  def collection
+    ::Collection.find_by!(druid: collection_druid)
+  rescue ActiveRecord::RecordNotFound
+    raise Error, "Collection not found: #{collection_druid}" if @raise
+
+    work.collection
+  end
+end
